src/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.7
bcrypt==3.1.7
boto3==1.14.47
botocore==1.17.47
cffi==1.14.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap3==12.1.0
django-braces==1.14.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-storages==1.9.1
docutils==0.15.2
Faker==4.1.0
gunicorn==20.0.4
jmespath==0.10.0
misaka==2.1.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
